Extract out gRPC tracing interceptor
