+++ rust/log/Cargo.toml
@@ -20,4 +20,4 @@ chroma-config = { workspace = true }
 chroma-error = { workspace = true }
 chroma-segment = { workspace = true }
 chroma-types = { workspace = true }
-
+chroma-tracing = { workspace = true, features = ["grpc"] }
