+++ rust/log/src/grpc_log.rs
@@ -1,5 +1,4 @@
 use super::config::LogConfig;
-use crate::tracing::client_interceptor;
 use crate::types::{
     CollectionInfo, GetCollectionsWithNewDataError, PullLogsError, UpdateCollectionLogOffsetError,
 };
@@ -7,6 +6,7 @@ use crate::PushLogsError;
 use async_trait::async_trait;
 use chroma_config::Configurable;
 use chroma_error::{ChromaError, ErrorCodes};
+use chroma_tracing::grpc_client_interceptor;
 use chroma_types::chroma_proto::log_service_client::LogServiceClient;
 use chroma_types::chroma_proto::{self};
 use chroma_types::{CollectionUuid, LogRecord, OperationRecord, RecordConversionError};
@@ -81,7 +81,7 @@ impl Configurable<LogConfig> for GrpcLog {
                                 tonic::transport::Channel,
                                 fn(Request<()>) -> Result<Request<()>, Status>,
                             >,
-                        > = LogServiceClient::with_interceptor(client, client_interceptor);
+                        > = LogServiceClient::with_interceptor(client, grpc_client_interceptor);
                         return Ok(GrpcLog::new(channel));
                     }
                     Err(e) => {
