+++ rust/sysdb/Cargo.toml
@@ -27,3 +27,4 @@ parking_lot = { workspace = true }
 chroma-config = { workspace = true }
 chroma-error = { workspace = true }
 chroma-types = { workspace = true }
+chroma-tracing = { workspace = true, features = ["grpc"] }
