+++ rust/sysdb/src/lib.rs
@@ -2,7 +2,6 @@ pub mod config;
 #[allow(clippy::module_inception)]
 pub mod sysdb;
 pub mod test_sysdb;
-mod util;
 pub use config::*;
 pub use sysdb::*;
 pub use test_sysdb::*;
