+++ rust/tracing/Cargo.toml
@@ -14,3 +14,6 @@ tracing-opentelemetry = { workspace = true }
 tracing-subscriber = { workspace = true }
 tracing.workspace = true
 tonic.workspace = true
+
+[features]
+grpc = []
